% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstringr-package.R
\docType{package}
\name{cstringr-package}
\alias{cstringr-package}
\title{cstringr}
\description{
Low level C string manipulation functions intended for use from C code in
other R packages.  Some functions are exposed via R for testing purposes, but
there is no point in using those since base R already implements all of the
functionality in this package.
}
