% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{cstringr-package}
\alias{cstringr-package}
\title{cstringr}
\description{
Low level C string manipulation functions.  This is the R interface that is
intended solely for testing since R already implements this functionality
and more.  The main use of the package is for the C functions that are
\code{R_RegisterCCallable}'d.
}

